feat: stock in history UI with mock data
