+++ components/inventory/stock-management/StockInHistory.tsx
@@ -1,0 +1,105 @@
+import { Button, Card, Flex, Typography } from "antd";
+import React from "react";
+
+export interface StockInHistoryProps {
+  id: number;
+  product_name: string;
+  stock: number;
+  product_sku: string;
+  warehouse_name: string;
+  invoice_number: string;
+  date: string;
+}
+
+interface StockInHistoryInterface {
+  items: StockInHistoryProps[];
+}
+
+const StockInHistory = ({ items }: StockInHistoryInterface) => {
+  return (
+    <Card
+      variant="outlined"
+      bodyStyle={{ paddingTop: 0, paddingLeft: 0, paddingRight: 0 }}
+      title={
+        <Flex
+          style={{
+            justifyContent: "space-between",
+            alignItems: "center",
+          }}
+        >
+          <Typography.Text
+            style={{ color: "#00000073", fontSize: 16, fontWeight: 500 }}
+          >
+            Recent Stock In
+          </Typography.Text>
+          <Button type="link" style={{ fontSize: 12, fontWeight: 400 }}>
+            View All Stock In History
+          </Button>
+        </Flex>
+      }
+    >
+      {items.map((item, index) => (
+        <div
+          key={item.id}
+          style={{
+            padding: "12px 24px",
+            borderBottom:
+              index !== items.length - 1 ? "1px solid #f0f0f0" : "none",
+          }}
+        >
+          <Flex
+            style={{
+              justifyContent: "space-between",
+              alignItems: "center",
+              // marginBottom: 4,
+            }}
+          >
+            <Typography.Text
+              style={{ color: "#000000D9", fontSize: 16, fontWeight: 500 }}
+            >
+              {item.product_name}
+            </Typography.Text>
+            <Typography.Text
+              style={{ color: "#52C41A", fontSize: 16, fontWeight: 500 }}
+            >
+              +{item.stock}
+            </Typography.Text>
+          </Flex>
+
+          <Flex
+            style={{
+              justifyContent: "space-between",
+              marginBottom: 2,
+            }}
+          >
+            <Typography.Text
+              style={{ color: "#00000073", fontSize: 14, fontWeight: 400 }}
+            >
+              {item.product_sku}
+            </Typography.Text>
+            <Typography.Text
+              style={{ color: "#00000073", fontSize: 14, fontWeight: 400 }}
+            >
+              {item.warehouse_name}
+            </Typography.Text>
+          </Flex>
+
+          <Flex style={{ justifyContent: "space-between" }}>
+            <Typography.Text
+              style={{ color: "#00000073", fontSize: 14, fontWeight: 400 }}
+            >
+              {item.invoice_number}
+            </Typography.Text>
+            <Typography.Text
+              style={{ color: "#00000073", fontSize: 14, fontWeight: 400 }}
+            >
+              {item.date}
+            </Typography.Text>
+          </Flex>
+        </div>
+      ))}
+    </Card>
+  );
+};
+
+export default StockInHistory;
