+++ components/inventory/stock-management/StockOut.tsx
